requirements.txt
google-generativeai 
python-dotenv
editdistance
chromadb
langchain
streamlit